+++ src/main/java/kitchenpos/exception/BadRequestException.java
@@ -1,0 +1,6 @@
+package kitchenpos.exception;
+
+public class BadRequestException extends RuntimeException {
+    public BadRequestException(String message) { super(message); }
+}
+
