+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,7 +1,8 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.ordertable.domain.OrderTable;
+
 import kitchenpos.exception.BadRequestException;
+import kitchenpos.ordertable.domain.OrderTable;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
